Change type to size_t so that 64 bit platforms dont whine about data loss when casting
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@5303 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/hashcomp.cpp
@@ -86,7 +86,7 @@ size_t nspace::hash<string>::operator()(const string &s) const
 	 * only with *x replaced with lowermap[*x].
 	 * This avoids a copy to use hash<const char*>
 	 */
-	register unsigned long t = 0;
+	register size_t t = 0;
 	for (std::string::const_iterator x = s.begin(); x != s.end(); x++) /* ++x not x++, so we don't hash the \0 */
 		t = 5 * t + lowermap[(unsigned char)*x];
 	return size_t(t);
